Move IntSurfLine to its own file
